requirements_windows.txt
ultralytics==8.1.18
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
pyinstaller==6.13.0 